requirements.txt
fastapi
uvicorn
pytesseract
Pillow
sentence-transformers
faiss-cpu
openai
python-dotenv
numpy
requests
beautifulsoup4
playwright
pydantic
